Cleanup some of the code by moving some of the features into the api module.
