+++ src/welcome.rs
@@ -1,0 +1,68 @@
+use crate::{
+    api,
+    commands::Args,
+    schema::{messages, users},
+    db::DB,
+    Result,
+};
+use diesel::prelude::*;
+use serenity::model::prelude::*;
+
+/// Write the welcome message to the welcome channel.  
+pub(crate) fn post_message(args: Args) -> Result {
+    use std::str::FromStr;
+
+    const WELCOME_BILLBOARD: &'static str = "By participating in this community, you agree to follow the Rust Code of Conduct, as linked below. Please click the :white_check_mark: below to acknowledge and gain access to the channels.
+
+  https://www.rust-lang.org/policies/code-of-conduct  ";
+
+    if api::is_mod(&args)? {
+        let channel_name = &args
+            .params
+            .get("channel")
+            .ok_or("unable to retrieve channel param")?;
+
+        let channel_id = ChannelId::from_str(channel_name)?;
+        info!("Posting welcome message");
+        let message = channel_id.say(&args.cx, WELCOME_BILLBOARD)?;
+        let bot_id = &message.author.id;
+
+        let conn = DB.get()?;
+
+        let _ = conn
+            .build_transaction()
+            .read_write()
+            .run::<_, Box<dyn std::error::Error>, _>(|| {
+                let message_id = message.id.0.to_string();
+                let channel_id = channel_id.0.to_string();
+
+                diesel::insert_into(messages::table)
+                    .values((
+                        messages::name.eq("welcome"),
+                        messages::message.eq(&message_id),
+                        messages::channel.eq(&channel_id),
+                    ))
+                    .on_conflict(messages::name)
+                    .do_update()
+                    .set((
+                        messages::message.eq(&message_id),
+                        messages::channel.eq(&channel_id),
+                    ))
+                    .execute(&conn)?;
+
+                let user_id = &bot_id.to_string();
+
+                diesel::insert_into(users::table)
+                    .values((users::user_id.eq(user_id), users::name.eq("me")))
+                    .on_conflict(users::name)
+                    .do_update()
+                    .set((users::name.eq("me"), users::user_id.eq(user_id)))
+                    .execute(&conn)?;
+                Ok(())
+            })?;
+
+        let white_check_mark = ReactionType::from("✅");
+        message.react(args.cx, white_check_mark)?;
+    }
+    Ok(())
+}
